Added an Update Result feature to separate adding registrations and adding results. Now I am going to start working on alerting the administrator if the event is over or not by seeing if all the students registered for an event have a result inputted or not

+++ db/define_schema.rb
@@ -21,12 +21,12 @@ DB.define_column("students", "college_id", "integer")
 
 DB.define_table("activities")
 DB.define_column("activities", "event_name", "string")
-DB.define_column("activities", "max_possible_score_for_activity", "integer")
+DB.define_column("activities", "max_possible_score_for_activity", "float")
 
 DB.define_table("results")
 DB.define_column("results", "student_id", "integer")
 DB.define_column("results", "activity_id", "integer")
-DB.define_column("results", "student_score", "integer")
+DB.define_column("results", "student_score", "float")
 
 DB.define_table("colleges")
 DB.define_column("colleges", "college_name", "string")
